refactor StdOutput to own widget
