+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+[v0.13.0] 2024-11-27
+-------------------
+
+-   refactor StdOutput to widget
+-   
+-   
+
 [v0.12.0] 2024-11-27
 -------------------
