+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "ck-cracker"
 include = ["src/**/*", "README.md", "CHANGELOG.md","LISCENSE", "fonts/**/*", "Cargo.toml"]
 default-run = "ck"
-version = "0.12.0"
+version = "0.13.0"
 edition = "2021"
 license = "GPL-3.0"
 reademe = "README.md"
